requirements.txt
langgraph
transformers
torch
langchain
python-multipart
beautifulsoup4
pypdf
pandas
requests